Slightly refactor.
[grodvi, grops, xditview, libbib, libgroff, eqn, pre-grohtml, preconv,
refer, troff, tfmtodit]: Store return value of `strlen()` in a `size_t`
(or `ptrdiff_t` if we're performing subtraction on it) instead of an
`int`.  Change temporaries (loop indices and similar) to use the same
types.

* src/devices/grodvi/dvi.cpp (dvi_printer::do_special):
* src/devices/grops/ps.cpp (ps_output::comment_arg):
  (ps_output::put_string, ps_output::put_number, ps_output::put_float)
  (ps_output::put_color):
* src/devices/grops/ps.cpp (class ps_output) <put_string, col>:
  <max_line_length>:
* src/devices/xditview/Dvi.c (ClassInitialize):
* src/include/symbol.h (class symbol) <block_size>:
* src/libs/libbib/index.cpp (index_search_item::munge_filename):
  (index_search_item::search_item):
* src/libs/libgroff/fontfile.cpp (font::open_file):
* src/libs/libgroff/string.cpp (string::operator=, string::operator+):
* src/libs/libgroff/symbol.cpp (symbol::symbol, symbol::block_size):
* src/preproc/eqn/delim.cpp (define_extensible_string):
* src/preproc/html/pre-html.cpp (char_buffer::can_see):
* src/preproc/html/pushback.cpp (pushBackBuffer::isString):
* src/preproc/preconv/preconv.cpp (emacs2mine):
* src/preproc/refer/label.ypp (format_expr::evaluate):
* src/roff/troff/input.cpp (make_temp_iterator, pipe_source):
* src/utils/tfmtodit/tfmtodit.cpp (main): Do it.

* src/preproc/html/pushback.cpp (pushBackBuffer::isString): Drop `while`
  loop that is now tautologically true due to comparison of unsigned
  types.

Also update editor aid comments; drop old style Emacs file-local
variable setting.

+++ src/devices/grops/ps.cpp
@@ -169,7 +169,7 @@ ps_output &ps_output::end_comment()
 
 ps_output &ps_output::comment_arg(const char *s)
 {
-  int len = strlen(s);
+  size_t len = strlen(s);
   if (col + len + 1 > max_line_length) {
     putc('\n', fp);
     fputs("%%+", fp);
@@ -200,10 +200,10 @@ ps_output &ps_output::put_delimiter(char c)
   return *this;
 }
 
-ps_output &ps_output::put_string(const char *s, int n)
+ps_output &ps_output::put_string(const char *s, size_t n)
 {
-  int len = 0;
-  int i;
+  size_t len = 0;
+  size_t i;
   for (i = 0; i < n; i++) {
     char c = s[i];
     if (is_ascii(c) && csprint(c)) {
@@ -290,7 +290,7 @@ ps_output &ps_output::put_number(int n)
 {
   char buf[1 + INT_DIGITS + 1];
   sprintf(buf, "%d", n);
-  int len = strlen(buf);
+  size_t len = strlen(buf);
   if (col > 0 && col + len + need_space > max_line_length) {
     putc('\n', fp);
     col = 0;
@@ -309,7 +309,7 @@ ps_output &ps_output::put_number(int n)
 ps_output &ps_output::put_fix_number(int i)
 {
   const char *p = if_to_a(i, fixed_point);
-  int len = strlen(p);
+  size_t len = strlen(p);
   if (col > 0 && col + len + need_space > max_line_length) {
     putc('\n', fp);
     col = 0;
@@ -329,7 +329,7 @@ ps_output &ps_output::put_float(double d)
 {
   char buf[128];
   sprintf(buf, "%.4f", d);
-  int last = strlen(buf) - 1;
+  ptrdiff_t last = strlen(buf) - 1;
   while (buf[last] == '0')
     last--;
   if (buf[last] == '.')
@@ -352,7 +352,7 @@ ps_output &ps_output::put_float(double d)
 
 ps_output &ps_output::put_symbol(const char *s)
 {
-  int len = strlen(s);
+  size_t len = strlen(s);
   if (col > 0 && col + len + need_space > max_line_length) {
     putc('\n', fp);
     col = 0;
@@ -372,7 +372,7 @@ ps_output &ps_output::put_color(unsigned int c)
 {
   char buf[128];
   sprintf(buf, "%.3g", double(c) / double(color::MAX_COLOR_VAL));
-  int len = strlen(buf);
+  size_t len = strlen(buf);
   if (col > 0 && col + len + need_space > max_line_length) {
     putc('\n', fp);
     col = 0;
@@ -390,7 +390,7 @@ ps_output &ps_output::put_color(unsigned int c)
 
 ps_output &ps_output::put_literal_symbol(const char *s)
 {
-  int len = strlen(s);
+  size_t len = strlen(s);
   if (col > 0 && col + len + 1 > max_line_length) {
     putc('\n', fp);
     col = 0;
