+++ src/libs/libgroff/fontfile.cpp
@@ -62,9 +62,10 @@ FILE *font::open_file(const char *nm, char **pathp)
   // Do not traverse user-specified directories; Savannah #61424.
   if (0 /* nullptr */ == strchr(nm, '/')) {
     // Allocate enough for nm + device + 'dev' '/' '\0'.
-    int expected_size = strlen(nm) + strlen(device) + 5;
+    size_t expected_size = strlen(nm) + strlen(device) + 5;
     char *filename = new char[expected_size];
-    const int actual_size = sprintf(filename, "dev%s/%s", device, nm);
+    const size_t actual_size = sprintf(filename, "dev%s/%s", device,
+				       nm);
     expected_size--; // sprintf() doesn't count the null terminator.
     if (actual_size == expected_size)
       fp = font_path.open_file(filename, pathp);
