+++ src/utils/tfmtodit/tfmtodit.cpp
@@ -764,7 +764,7 @@ int main(int argc, char **argv)
   if (special_flag)
     fputs("special\n", stdout);
   char *internal_name = strsave(argv[optind]);
-  int len = strlen(internal_name);
+  size_t len = strlen(internal_name);
   if (len > 4 && strcmp(internal_name + len - 4, ".tfm") == 0)
     internal_name[len - 4] = '\0';
   // DIR_SEPS[] are possible directory separator characters, see
